+++ test/projects/standard/server/lifecycle/on-render-after-data-to-store.js
@@ -1,4 +1,11 @@
-export default async (/*o*/) => {
-    if (__DEV__) console.log('server lifecycle: onRenderAfterDataToStore')
+export default async ({ ctx }) => {
+    if (__DEV__) {
+        console.log('server lifecycle: onRenderAfterDataToStore')
+        if (/^\/delayed(\/|$)/.test(ctx.path)) {
+            await new Promise(resolve =>
+                setTimeout(resolve, 1500)
+            )
+        }
+    }
     // if (__DEV__) console.log(o)
 }